+++ src/main/java/org/mySite/user/service/UserService.java
@@ -1,11 +1,13 @@
 package org.mySite.user.service;
 
 import lombok.RequiredArgsConstructor;
+import org.mySite.user.SpringSecurity.PrincipalDetails;
 import org.mySite.user.domain.User;
 import org.mySite.user.dto.JoinRequest;
 import org.mySite.user.dto.LoginRequest;
 import org.mySite.user.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -19,43 +21,9 @@ public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public User getLoginUser(Long loginRequest){
-        User user = userRepository.findById(loginRequest)
+    public User getLoginUser(Long id){
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("UserId not found"));
-
-        return user;
-    }
-
-    // ------ 회원가입 --------
-
-    // loginId 중복 체크
-    public boolean checkLoginId(String loginId) {
-        // 1건이라도 있으면 바로 스캔 종료 후 true 리턴
-        return userRepository.existsByLoginId(loginId);
-    }
-
-    // nickname 중복 체크
-    public boolean checkNickname(String nickname) {
-        return userRepository.existsByNickname(nickname);
-    }
-
-    // JoinRequest을 입력 받아 User로 변환
-    public void join(JoinRequest request) {
-        userRepository.save(request.toEntity());
-    }
-
-
-    // ------로그인-------
-    public User login(LoginRequest request){
-
-        // 일치하는 User 찾기
-        User user = userRepository.findByLoginId(request.getLoginId())
-                .orElseThrow(() -> new IllegalArgumentException("User not found"));
-
-        // 찾은 User 비밀번호와 입력된 비밀번호 일치하는지 확인
-        if(!user.getPassword().equals(request.getPassword())) {
-            return null;
-        }
 
         return user;
     }
@@ -72,20 +40,30 @@ public class UserService {
         return optionalUser.get();
     }
 
-    // 회원가입 2
-    // 스프링 시큐리티에서 사용 + JWT에서 사용
-    // 비밀번호 암호화 사용
+    // ------ 회원가입 --------
+
+    // loginId 중복 체크
+    public boolean checkLoginId(String loginId) {
+        // 1건이라도 있으면 바로 스캔 종료 후 true 리턴
+        return userRepository.existsByLoginId(loginId);
+    }
+
+    // nickname 중복 체크
+    public boolean checkNickname(String nickname) {
+        return userRepository.existsByNickname(nickname);
+    }
+
+    // spring security + jwt 사용
+    // 비밀번호 암호화 저장
     public void join2(JoinRequest request){
         userRepository.save(request.toEntity(passwordEncoder.encode(request.getPassword())));
     }
 
-    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
-        User user = userRepository.findByLoginId(loginId)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        return org.springframework.security.core.userdetails.User.builder()
-                .username(user.getLoginId())
-                .password(user.getPassword())
-                .authorities(user.getRole().name())
-                .build();
-    }
+    // ------로그인-------
+//    @Override
+//    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
+//        User user = userRepository.findByLoginId(loginId)
+//                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+//        return new PrincipalDetails(user);
+//    }
 }
