+++ src/main/resources/templates/main.html
@@ -11,10 +11,10 @@
 <br><br>
 <h3><a href="/session-login">세션 로그인</a>인뎁쇼!</h3>
 <br>
-<h3><a href="/security-login">Security 로그인</a></h3>
 <br>
 
 <hr>
+<h3><a href="/auth">로그인</a></h3>
 <h3>JWT 로그인은 POSTMAN 사용하면 됨</h3>
 <br>
 <hr>
